pyInference/inferenceNode/requirements.txt
torch==2.6.0
grpcio==1.70.0
grpcio-tools==1.70.0
transformers==4.47.1
protobuf==5.29.2
